+++ ausphin_special/views/crm_lead_views.xml
@@ -19,6 +19,10 @@
         <field name="model">crm.lead</field>
         <field name="inherit_id" ref="crm.crm_case_form_view_oppor"/>
         <field name="arch" type="xml">
+            <button name="toggle_active" position="after">
+                <button string="Move to Prev" type="object" name="action_move_to_prev_stage"/>
+                <button string="Move to Next" type="object" name="action_move_to_next_stage"/>
+            </button>
             <xpath expr="//label[@for='planned_revenue']/.." position="attributes">
                 <attribute name="invisible">1</attribute>
             </xpath>
