--- crm_move_stage_button/models/__init__.py
@@ -1,2 +1,0 @@
-# -*- coding: utf-8 -*-
-from . import crm_lead
