--- crm_move_stage_button/views/crm_lead_views.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <record id="crm_lead_view_form_inherit" model="ir.ui.view">
-        <field name="name">crm.lead.view.form.inherit</field>
-        <field name="model">crm.lead</field>
-        <field name="inherit_id" ref="crm.crm_case_form_view_oppor"/>
-        <field name="arch" type="xml">
-            <button name="toggle_active" position="after">
-                <button string="Move to Prev" type="object" name="action_move_to_prev_stage"/>
-                <button string="Move to Next" type="object" name="action_move_to_next_stage"/>
-            </button>
-        </field>
-    </record>
-</odoo>
